Moved disabler to interpreterbase.
