+++ mesonbuild/interpreter.py
@@ -27,7 +27,7 @@ from .dependencies import InternalDependency, Dependency, DependencyException
 from .interpreterbase import InterpreterBase
 from .interpreterbase import check_stringlist, noPosargs, noKwargs, stringArgs, permittedKwargs
 from .interpreterbase import InterpreterException, InvalidArguments, InvalidCode
-from .interpreterbase import InterpreterObject, MutableInterpreterObject
+from .interpreterbase import InterpreterObject, MutableInterpreterObject, Disabler
 from .modules import ModuleReturnValue
 
 import os, sys, shutil, uuid
@@ -1081,14 +1081,6 @@ ModuleState = namedtuple('ModuleState', [
     'data', 'headers', 'man', 'global_args', 'project_args', 'build_machine',
     'host_machine', 'target_machine'])
 
-class Disabler(InterpreterObject):
-    def __init__(self):
-        super().__init__()
-        self.methods.update({'found': self.found_method})
-
-    def found_method(self, args, kwargs):
-        return False
-
 class ModuleHolder(InterpreterObject, ObjectHolder):
     def __init__(self, modname, module, interpreter):
         InterpreterObject.__init__(self)
@@ -1097,8 +1089,6 @@ class ModuleHolder(InterpreterObject, ObjectHolder):
         self.interpreter = interpreter
 
     def method_call(self, method_name, args, kwargs):
-        if is_disabled(args, kwargs):
-            return Disabler()
         try:
             fn = getattr(self.held_object, method_name)
         except AttributeError:
@@ -1389,19 +1379,6 @@ permitted_kwargs = {'add_global_arguments': {'language'},
                     'vcs_tag': {'input', 'output', 'fallback', 'command', 'replace_string'},
                     }
 
-def is_disabled(args, kwargs):
-    for i in args:
-        if isinstance(i, Disabler):
-            return True
-    for i in kwargs.values():
-        if isinstance(i, Disabler):
-            return True
-        if isinstance(i, list):
-            for j in i:
-                if isinstance(j, Disabler):
-                    return True
-    return False
-
 
 class Interpreter(InterpreterBase):
 
@@ -1604,8 +1581,6 @@ class Interpreter(InterpreterBase):
     @noPosargs
     def func_declare_dependency(self, node, args, kwargs):
         version = kwargs.get('version', self.project_version)
-        if is_disabled(args, kwargs):
-            return Disabler()
         if not isinstance(version, str):
             raise InterpreterException('Version must be a string.')
         incs = extract_as_list(kwargs, 'include_directories', unholder=True)
@@ -2156,20 +2131,6 @@ to directly access options of other subprojects.''')
                     break
         return identifier, cached_dep
 
-    @permittedKwargs(permitted_kwargs['dependency'])
-    def is_disabled(self, args, kwargs):
-        for i in args:
-            if isinstance(i, DisablerHolder):
-                return True
-        for i in kwargs.values():
-            if isinstance(i, DisablerHolder):
-                return True
-            if isinstance(i, list):
-                for j in i:
-                    if isinstance(j, DisablerHolder):
-                        return True
-        return False
-
     def func_dependency(self, node, args, kwargs):
         self.validate_arguments(args, 1, [str])
         if 'required' in kwargs and 'options' in kwargs:
@@ -2441,8 +2402,6 @@ to directly access options of other subprojects.''')
 
     @permittedKwargs(permitted_kwargs['test'])
     def func_test(self, node, args, kwargs):
-        if is_disabled(args, kwargs):
-            return Disabler()
         self.add_test(node, args, kwargs, True)
 
     def unpack_env_kwarg(self, kwargs):
@@ -2967,8 +2926,6 @@ different subdirectory.
             self.coredata.target_guids[idname] = str(uuid.uuid4()).upper()
 
     def build_target(self, node, args, kwargs, targetholder):
-        if is_disabled(args, kwargs):
-            return Disabler()
         if not args:
             raise InterpreterException('Target does not have a name.')
         name = args[0]
