+++ test cases/common/161 dolphin option/badgenerator.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python3
+
+print('I am a bad generator. I do not work.')
