+++ test cases/common/161 dolphin option/meson.build
@@ -6,3 +6,10 @@ exe = executable('encprog', 'encprog.c',
   dependencies : x264_dep)
 test('enc test', exe)
 
+gs = custom_target('somec',
+  input : 'proggie.c.in',
+  output : 'proggie.c',
+  command : [exe, '@INPUT@', '@OUTPUT@'],
+  )
+
+test('Generated test', executable('proggie', gs))
